feat(tui): add unified async acceptance predicate
